Separate ssl_vae + Fix SslVae training
