+++ examples/vae_ssl_mnist.py
@@ -1,5 +1,5 @@
 from functools import partial
-from pytorch_autoencoders.models import vae
+from pytorch_autoencoders.models import ssl_vae
 from pytorch_autoencoders.config import Config
 from pytorch_autoencoders import inference_helper, train_helper
 import torch
@@ -11,15 +11,14 @@ from torchvision.datasets import MNIST
 def train() -> None:
     config = Config()
     config.optim = partial(Adam, lr=0.001, weight_decay=1e-5)
+    config.batch_size = 200
     data = MNIST("/tmp/mnist/train", download=True, transform=ToTensor())
-    config.criterion = vae.bernoulli_loss
-    ae = vae.VariationalAutoEncoder(torch.Size((28, 28)), config)
-    result = train_helper.train(ae, config, data, train_helper.vae_logfn)
+    config.criterion = ssl_vae.labeled_bernoulli_loss
+    ae = ssl_vae.VAESslM2(torch.Size((28, 28)), config, nlabels=10)
+    result = train_helper.train_ss(ae, config, data, train_helper.vae_logfn)
     result.to_csv("vae_mnist_result.csv")
-    ae.save("vae.pth")
+    ae.save("vae_ssl.pth")
     data = MNIST("/tmp/mnist/test", download=True, train=False, transform=ToTensor())
-    inference_helper.show_decoded_images(ae, config, data)
-    inference_helper.show_feature_map(ae, config, data)
 
 
 if __name__ == "__main__":
